+++ config/config.exs
@@ -1,3 +1,4 @@
 import Config
 
 # Add your configuration here
+config :logger, level: :info